AdaCompress-master/requirements1.txt
Keras
#res_manager
numpy
tensorflow_gpu
boto3
opencv_python
matplotlib
#Algorithmia
Pillow
